KBHandler/requirements.txt
farm-haystack==1.10.0
transformers==4.21.2